projects/mdx-chatbot/requirements.txt
langchain>=0.0.267
langchain-community>=0.0.1
langchain-ollama>=0.0.1
python-frontmatter>=1.0.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
python-dotenv>=1.0.0